requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
ImageHash==4.3.1
Pillow==9.4.0
requests==2.28.1
